+++ Qa.Core/Structure/DType.cs
@@ -3,7 +3,7 @@
     public enum DType
     {
         None = 0,
-        Float = 1,
+        Double = 1,
         Int = 2,
         Money = 3,
         Percent = 4
